+++ relay-event-schema/src/protocol/span/convert.rs
@@ -50,6 +50,11 @@ impl From<&Event> for Span {
             trace_id: trace.map(|c| c.trace_id.clone()).unwrap_or_default(),
             segment_id: trace.map(|c| c.span_id.clone()).unwrap_or_default(),
             is_segment: true.into(),
+            // NB: Technically, this span may not be an actual remote span if this is a child
+            // transaction created within the same service as its parent. We still set `is_remote`
+            // as the best proxy to ensure this span will be detected as a segment by the spans
+            // pipeline.
+            is_remote: true.into(),
             status: trace.map(|c| c.status.clone()).unwrap_or_default(),
             description: transaction.clone(),
             tags: tags.clone().map_value(|t| t.into()),
@@ -146,6 +151,7 @@ mod tests {
                 "fa90fdead5f74052",
             ),
             is_segment: true,
+            is_remote: true,
             status: Ok,
             description: "my 1st transaction",
             tags: ~,
@@ -214,6 +220,8 @@ mod tests {
                 messaging_message_receive_latency: ~,
                 messaging_message_body_size: ~,
                 messaging_message_id: ~,
+                messaging_operation_name: ~,
+                messaging_operation_type: ~,
                 user_agent_original: ~,
                 url_full: ~,
                 client_address: ~,