+++ relay-server/src/metrics_extraction/event.rs
@@ -1166,7 +1166,9 @@ mod tests {
                         "messaging.message.receive.latency": 100,
                         "messaging.message.retry.count": 2,
                         "messaging.message.body.size": 1000,
-                        "messaging.message.id": "abc123"
+                        "messaging.message.id": "abc123",
+                        "messaging.operation.name": "publish",
+                        "messaging.operation.type": "create"
                     }
                 },
                 {
@@ -1182,7 +1184,9 @@ mod tests {
                         "messaging.message.receive.latency": 100,
                         "messaging.message.retry.count": 2,
                         "messaging.message.body.size": 1000,
-                        "messaging.message.id": "abc123"
+                        "messaging.message.id": "abc123",
+                        "messaging.operation.name": "publish",
+                        "messaging.operation.type": "create"
                     }
                 },
                 {
@@ -1198,7 +1202,9 @@ mod tests {
                         "messaging.message.receive.latency": 100,
                         "messaging.message.retry.count": 2,
                         "messaging.message.body.size": 1000,
-                        "messaging.message.id": "abc123"
+                        "messaging.message.id": "abc123",
+                        "messaging.operation.name": "publish",
+                        "messaging.operation.type": "create"
                     }
                 },
                 {
@@ -1214,7 +1220,9 @@ mod tests {
                         "messaging.message.receive.latency": 100,
                         "messaging.message.retry.count": 2,
                         "messaging.message.body.size": 1000,
-                        "messaging.message.id": "abc123"
+                        "messaging.message.id": "abc123",
+                        "messaging.operation.name": "publish",
+                        "messaging.operation.type": "create"
                     }
                 },
                 {
@@ -1812,7 +1820,7 @@ mod tests {
         );
 
         for mri in [
-            "d:transactions/measurements.cls@ratio",
+            "d:transactions/measurements.cls@none",
             "d:transactions/measurements.score.cls@ratio",
             "d:transactions/measurements.score.total@ratio",
             "d:transactions/measurements.score.weight.cls@ratio",
@@ -1852,7 +1860,7 @@ mod tests {
         );
 
         for mri in [
-            "d:transactions/measurements.lcp@ratio",
+            "d:transactions/measurements.lcp@millisecond",
             "d:transactions/measurements.score.lcp@ratio",
             "d:transactions/measurements.score.total@ratio",
             "d:transactions/measurements.score.weight.lcp@ratio",